Refactor GetMatches activity, GetPets activity, and related classes into packages
