+++ app/src/main/AndroidManifest.xml
@@ -68,7 +68,7 @@
                 <category android:name="android.intent.category.LAUNCHER"/>
             </intent-filter>
         </activity>
-        <activity android:name=".GetMatchesActivity"
+        <activity android:name=".ui.get_matches.GetMatchesActivity"
             android:exported="true"
             android:label="">
             <intent-filter>
@@ -77,7 +77,7 @@
                 <category android:name="android.intent.category.LAUNCHER"/>
             </intent-filter>
         </activity>
-        <activity android:name=".GetPetsActivity"
+        <activity android:name=".ui.get_pets.GetPetsActivity"
             android:exported="true"
             android:label="">
             <intent-filter>
