+++ app/src/main/java/com/example/cupetfrontend/ui/get_matches/GetMatchesResult.java
@@ -1,4 +1,4 @@
-package com.example.cupetfrontend;
+package com.example.cupetfrontend.ui.get_matches;
 
 import androidx.annotation.Nullable;
