+++ app/src/main/java/com/example/cupetfrontend/ui/get_matches/RecyclerViewAdapter.java
@@ -1,4 +1,4 @@
-package com.example.cupetfrontend;
+package com.example.cupetfrontend.ui.get_matches;
 
 import android.app.AlertDialog;
 import android.content.Context;
@@ -17,6 +17,7 @@ import androidx.recyclerview.widget.RecyclerView;
 
 import com.bumptech.glide.Glide;
 import com.bumptech.glide.request.RequestOptions;
+import com.example.cupetfrontend.R;
 import com.example.cupetfrontend.data.model.PetModel;
 
 import java.util.ArrayList;
