+++ app/src/main/res/layout/activity_view_pets.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".GetPetsActivity"
+    tools:context=".ui.get_pets.GetPetsActivity"
     android:layout_marginStart="5dp"
     android:paddingStart="10dp">
